requirements.txt
async-generator==1.10
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
exceptiongroup==1.1.1
Flask==2.3.2
Flask-Dance==7.0.0
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
google-auth==2.19.0
google-auth-oauthlib==1.0.0
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
oauthlib==3.2.2
outcome==1.2.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.9.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==2.0.15
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.6.2
urllib3==1.26.16
URLObject==2.4.3
Werkzeug==2.3.4
wsproto==1.2.0
